Centralized config framework
The config framework consist of global config state and model specific config state.
There is a preconfig phase where all the command line options and configurations are
defined and this is followed by command line parsing. After parsing the command line options,
there is a postconfig phase where the configuration is validated before building the model.

Both preconfig and postconfig phases comprises of a global run and a model specific run.

This changeset deprecates the model specific Configure() method.