Configurable Disk Latency function framework
This changeset implements a framework for specifying a latency function in
the command line. This is done as an interface DiskLatencySimulator. This
interface can be implemented in each model which needs a specific form latency function.
The interface has a name and this name can be specified in command line.

disk.go implements a generic latency function called latency-nil